semantic_news_navigator/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
umap-learn>=0.5.4
plotly>=5.18.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
